Робота з бібліотеками
Створено нову директорію lib, щоб тримати класи fiction.rb і book.rb. Також перенесено файли dolphin.rb і monkey.rb у директорію lib і змінено файл literature.rb щоб використати require між файлами.
