+++ ruby_classes/lib/dolphin.rb
