+++ ruby_classes/lib/monkey.rb
